+++ config/bootstrap.php
@@ -24,15 +24,14 @@ if (!Configure::check('Thumber.target')) {
 
 //Checks for driver
 $driver = Configure::read('Thumber.driver');
-
 if (!in_array($driver, ['imagick', 'gd'])) {
-    trigger_error(sprintf('The driver %s is not supported', $driver), E_USER_ERROR);
+    trigger_error(sprintf('The driver `%s` is not supported', $driver), E_USER_ERROR);
 }
 
 //Checks for target directory
 $target = Configure::read('Thumber.target');
-safe_mkdir($target, 0777, true);
+@mkdir($target, 0777, true);
 
 if (!is_writeable($target)) {
-    trigger_error(sprintf('Directory %s not writeable', $target), E_USER_ERROR);
+    trigger_error(sprintf('Directory `%s` not writeable', $target), E_USER_ERROR);
 }