+++ src/Command/ClearAllCommand.php
@@ -14,10 +14,10 @@
 namespace Thumber\Command;
 
 use Cake\Console\Arguments;
-use Cake\Console\Command;
 use Cake\Console\ConsoleIo;
 use Cake\Console\ConsoleOptionParser;
 use Exception;
+use MeTools\Console\Command;
 use Thumber\Utility\ThumbManager;
 
 /**
@@ -26,15 +26,28 @@ use Thumber\Utility\ThumbManager;
 class ClearAllCommand extends Command
 {
     /**
+     * @var \Thumber\Utility\ThumbManager
+     */
+    public $ThumbManager;
+
+    /**
+     * Hook method invoked by CakePHP when a command is about to be executed
+     * @return void
+     * @uses $ThumbManager
+     */
+    public function initialize()
+    {
+        $this->ThumbManager = $this->ThumbManager ?: new ThumbManager;
+    }
+
+    /**
      * Hook method for defining this command's option parser
      * @param ConsoleOptionParser $parser The parser to be defined
      * @return ConsoleOptionParser
      */
     protected function buildOptionParser(ConsoleOptionParser $parser)
     {
-        $parser->setDescription(__d('thumber', 'Clears all thumbnails'));
-
-        return $parser;
+        return $parser->setDescription(__d('thumber', 'Clears all thumbnails'));
     }
 
     /**
@@ -42,12 +55,12 @@ class ClearAllCommand extends Command
      * @param Arguments $args The command arguments
      * @param ConsoleIo $io The console io
      * @return null|int The exit code or null for success
-     * @uses ThumbManager::clearAll()
+     * @uses $ThumbManager
      */
     public function execute(Arguments $args, ConsoleIo $io)
     {
         try {
-            $count = (new ThumbManager)->clearAll();
+            $count = $this->ThumbManager->clearAll();
         } catch (Exception $e) {
             $io->err(__d('thumber', 'Error deleting thumbnails'));
             $this->abort();