+++ src/Command/ClearCommand.php
@@ -14,10 +14,10 @@
 namespace Thumber\Command;
 
 use Cake\Console\Arguments;
-use Cake\Console\Command;
 use Cake\Console\ConsoleIo;
 use Cake\Console\ConsoleOptionParser;
 use Exception;
+use MeTools\Console\Command;
 use Thumber\Utility\ThumbManager;
 
 /**
@@ -26,19 +26,32 @@ use Thumber\Utility\ThumbManager;
 class ClearCommand extends Command
 {
     /**
+     * @var \Thumber\Utility\ThumbManager
+     */
+    public $ThumbManager;
+
+    /**
+     * Hook method invoked by CakePHP when a command is about to be executed
+     * @return void
+     * @uses $ThumbManager
+     */
+    public function initialize()
+    {
+        $this->ThumbManager = $this->ThumbManager ?: new ThumbManager;
+    }
+
+    /**
      * Hook method for defining this command's option parser
      * @param ConsoleOptionParser $parser The parser to be defined
      * @return ConsoleOptionParser
      */
     protected function buildOptionParser(ConsoleOptionParser $parser)
     {
-        $parser->setDescription(__d('thumber', 'Clears all thumbnails that have been generated from an image path'));
-        $parser->addArgument('path', [
-            'help' => __d('thumber', 'Path of the original image'),
-            'required' => true,
-        ]);
-
-        return $parser;
+        return $parser->setDescription(__d('thumber', 'Clears all thumbnails that have been generated from an image path'))
+            ->addArgument('path', [
+                'help' => __d('thumber', 'Path of the original image'),
+                'required' => true,
+            ]);
     }
 
     /**
@@ -46,12 +59,12 @@ class ClearCommand extends Command
      * @param Arguments $args The command arguments
      * @param ConsoleIo $io The console io
      * @return null|int The exit code or null for success
-     * @uses ThumbManager::clear()
+     * @uses $ThumbManager
      */
     public function execute(Arguments $args, ConsoleIo $io)
     {
         try {
-            $count = (new ThumbManager)->clear($args->getArgument('path'));
+            $count = $this->ThumbManager->clear($args->getArgument('path'));
         } catch (Exception $e) {
             $io->err(__d('thumber', 'Error deleting thumbnails'));
             $this->abort();